benchmarking/api-benchmarking-locust/pyproject.toml
[project]
name = "api-benchmarking-locust"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "locust>=2.41.3",
    "openai>=2.1.0",
    "python-dotenv>=1.1.1",
]
